Add files via upload
Marquis is building a collection of apps and utilities for Windows. He has an unfinished project
which he wants you to complete. This project is a small text-based role-playing game.

Marquis wants me to create a few custom character objects for his game. In particular, one
character for the player and two enemy characters the computer will control. He has written a
few methods for the characters and a sample battle scenario. I need to add the methods and
properties to the characters as well.